Two updates to allow building on Cray machines with intrinsic double precision (e.g. -fdefault-real-8 for GNU):
1) Instead of passing 1.0d-1, need to pass 1.0e-1_cvmix_r8 (some compilers will treat 1.0d-1 as a 16-byte real if you use real-8 by default)

2) To convert from integer values to cvmix_r8, use (1.0_cvmix_r8*int_val) instead of dble(int_val). Again, some compilers treat dble(int_val) as a conversion to 16-byte if you use real-8 by default.

Testing:

i) Built on hopper with -fdefault-real-8, verified I was getting same error Alistair mentioned seeing on the GFDL machine

ii) applied fix that is being committed below, verified that double-diff, shear, tidal, and background-mixing reg_tests don't change answers. (I should have saved the output from old reg tests run on hopper with GNU, but instead I compared to the output on my laptop.)

Note that only one updated file is in the shared directory, but I figured I would make the stand-alone driver consistent as well in case we decide to add a test for this issue when we start automating the test process.




git-svn-id: https://cvmix.googlecode.com/svn/trunk@133 ec264166-b127-864d-6cb2-e100125a5c76