Big update for KPP: added routine to compute boundary layer depth from bulk Richardson number values (on cell centers). Default is to use quadratic interpolation, but options for linear or cubic splines as well.
Other small changes: formatting in a couple of routines, addition of critical Richardson number (default = 0.3) and interpolation type to kpp_params data type.

The KPP driver currently computes the OBL in a single column based on the following setup: given a parameter hmix and a critical Richardson number, the bulk Richardson number is assumed to be 0 from the surface to depth of hmix, and then increasing linearly at a rate of 0.5*Ri_crit per meter (so the true OBL depth is hmix-2 meters deep, given the convention that depth is positive up). At each level, the Ri_bulk is computed to be the integral average over the entire level, and then values are interpolated between cell centers to determine where the boundary layer actually is.

From my testing, quadratic (the POP default) is better than using cubic splines, which is better than linear interpolation.



git-svn-id: https://cvmix.googlecode.com/svn/trunk@176 ec264166-b127-864d-6cb2-e100125a5c76