After swapping emails with Bill, I was able to worked the last of the kinks in the latest test for the KPP driver. Basically, the issue was that he was computing the buoyancy frequency at interfaces using buoyancy values at cell centers (and a centered-difference derivative operator). He was then averaging the interface values to get buoyancy frequency at the cell center.
As with the last change, this does not affect the code in the shared module, it's only a change in the stand-alone driver. It does, however, make me more confident that cvmix_kpp_compute_bulk_Richardson and cvmix_kpp_compute_unresolved_shear are correct. Hooray!

I still need to clean up the driver code - adding parameters to the namelist, generating output, etc... but that's not a high priority at this time. (I'll still create an issue tag so I don't forget to do it further down the line.)



git-svn-id: https://cvmix.googlecode.com/svn/trunk@238 ec264166-b127-864d-6cb2-e100125a5c76